applied refactoring already in main to make merge easier
